moved utils.py to common/utils.py
