+++ backend/python/app/services/graph_db/common/utils.py
